Move endpoint type from FFI to ION crate
